Work for 9/30/24 In Class
Created the front page for the Venice tour page. This page will showcase the 3 tours that our company offers. Today I added the first container for the webpage. I added animations for the text on the background image. The animation is a slide up animation. The font that I used was SF-Pro.