Index Update
In this update, a button was added to the showcase section of the landing page. This button includes a question and learn more, which brings the user to tours. I then began formatting and the color scheme of the experiences section of the landing page, which will include details about the different features of tours offered.