Index Further Formatting
In this commit, the index page was formatted even more. Regarding the experiences and why us pages specifically, the boxes were adjusted to be fluid for text, and the spacing between elements was changed to increase appearance. The page overall was also changed to include a media querry that adjusted for resizing.