Login/Create/SignOut
In this commit, I did a lot of work creating the login, create account, and sign out functionality. I added pages to sign in and create an account. I did all the javascript for signing in and creating accounts in addition to the javascript for each page which determines if a user is signed in and changes the buttons. I did tours.js to tell if a user is signed in and work in conjunction with the buttons to book a tour or direct the user to sign in. I customized the css for the new html pages and made them look nice and ready to deploy.